requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
python-multipart
requests
numpy>=1.22,<2.0
pandas
scikit-learn
shap
torch
torchvision

pillow
scikit-image
mediapipe==0.10.13
captum
google-auth
xgboost
joblib
tqdm
matplotlib
python-jose[cryptography]
passlib[bcrypt]
pydantic
gunicorn
opencv-python-headless

